feat: Add scripts to help with making fast database backups
Since the standard restore script wipes out the database and uses the
SQL backup file to do a database restore on startup, this can be
somewhat irritating and time-consuming. In development environments,
these "fast" backup/restore scripts can backup and restore the raw files
of the database, making it faster to "reset" a database to the point of
the backup because the database doesn't need to do a SQL restore.

However, this is not an approved form of backup. It should never be used
for any production work.