feat: Add `mdl` shortcut script
This script is in a `bin` directory, and if you add the `bin` directory
to your path, you can use `mdl` as a parent command for all the scripts,
similar to what git or docker CLI tools do.

For instance, instead of calling `./start.sh`, you could call:

mdl start

And you can call it from anywhere instead of needing to explicitly call
it from the moodle repo directory.