feat: Add `exec-sql` script for easily running SQL scripts in envs
Similar to the `cli` script for PHP scripts, the `exec-sql` script will
find the MariaDB container for an environment and execute either a SQL
statement or a SQL file.

If the received param is a file, it pipes in the file contents.
Otherwise, it treats it like a SQL statement to execute.