feat: Enhance start script to have a `--wait` parameter
If the `--wait` parameter is used, the start script will not exit until
the Moodle environment is fully started. This means not just that the
containers exist, but the Moodle "check" CLI task is returning success.

This can be used by any scripts that need to wait until the Moodle
environment is fully running before proceeding with other tasks.