refactor: Put config updates in a separate script
At the end of a restore, we update Moodle's `config.php` file with
settings that work in the Docker environment, including updates of
settings in the `.env` file. By separating this logic in a separate
`update-config.sh` script, it can be called by multiple scripts,
including doing these updates each time before we start an environment
with `start.sh`.