perf: Switch to Docker volumes for the Moodle database
Since the database can run faster in a Docker volume as opposed to a
bind mount, we're updating the stacks to use a Docker volume for the
database.

It's still simpler to keep the `src` and `data` directories as bind
mounts since we do a lot of manipulation of those files.

Scripts that worked with the `db` directory had to be updated to use
the volume accordingly. In the case of clearing the database, this just
involves deleting the volume. In the case of the fast DB backup/restore
scripts, they were updated to spin up a transient Alpine container to
just run the tar command on the database files.