ci: Update docker compose files to not use Swarm and work with Podman
To be compatible with both Docker and Podman, we need to adjust things.
The compose files now reference environments relative to the project
root, which also means we don't need symbolic links to the docker
compose files in each environment.

That also means we pull in the Moodle Environment name (`mname`) to both
provide a stack name and to define the path to the environment files.

The containers must also run in privileged mode.