build: Update docker compose files to use custom entrypoint
Instead of building our own Docker image that customized Bitnami's image, we just make the
necessary adjustments in the docker compose file. Until we have a need to incorporate more
customizations in the image itself, this is simpler.

Note that if entrypoint is customized, the command will be ignored. So, even though we're
using the default command, we must specify it since we're specifying a custom entrypoint.

Ref: https://docs.docker.com/reference/compose-file/services/#entrypoint