feat: Improve permission handling of restore and sync script
Previously, the restore and sync scripts would try to fix permissions after updating
files, and this was very imperfect. When assigning the docker user permission, which is
only needed on docker installations in Linux, it was not properly checking the ID of the
docker user. And it was trying to fix permissions of data/src directories, but it wasn't
doing it quite right according to guidelines (see ref).

Fix discovery of docker user, and properly assign permissions if docker user exists.

Removed fixing permissions for data/src directories, since those are now fixed by the
custom entrypoint every time the container starts.

Ref: https://docs.bitnami.com/general/how-to/migrate-moodle/