feat: Enhance `mdl cli` to list available commands if no cmd is provided
If you don't provide a command, `mdl cli` will query the environment for the list of
commands in the cli directory.

Also reworked the logic of checking that the container is running, since the command list
also requires the container to be running to function.

Lastly, fixed a "bug" where the '-h' or '--help' flag would display the `mdl cli` help
screen no matter what, even if you tried to pass the flag to a CLI command.