fix: Make readlink ignore additional flags in `script_name` function
The `script_name` function uses `ps` to get the full command that caused the script to get
executed. This may include flags. We want `readlink` to process all the output with word
splitting, but we don't want it to process additional flags in the process. So we tell it
to ignore any further flags with `--`.