feat: Add `mdl config` script to simplify Moodle configurations
The `mdl config` script is partly a shortcut to use the `mdl cli` script to run Moodle's
`cfg` admin CLI, similar to how `mdl resetpassword` is for the `reset_password` admin CLI.

However, it also will scan the .env file and apply all valid Moodle configurations that
are defined in it. Since Moodle configurations are always lowercase, there will never be
any overlap with other configurations.

By querying the `cfg` CLI for each key, it verifies that it will only attempt to set keys
that are valid configs.