fix: Improve `export-env.sh` to only look at uppercase keys
The .env file may contain lowercase keys that are used for different configurations, but
are still useful to place in the .env file. All true environment variables, which is what
this script is exporting, will be UPPERCASE keys.

So, we modified export-env.sh to only find lines that have UPPERCASE keys followed by an
equal (=) sign. This effectively handles omitting comment lines and blank lines too. This
improvement avoids cluttering memory with configurations that aren't meant to be loaded.

Technically: UPPERCASE, numeric, and underscore (_) are allowed characters for the keys.