feat: Add box support to `mdl list`
To accomplish this, "box" is added to the list of possible `type` values, although the
default of `backup fastdb` remains, so that querying the Box.com API does not unduly slow
down the list command.

Since adding "box" to the full list would irritatingly require setting the type flag of
`-t 'backup fastdb box`, we add an alternative method to just provide `-b` or `--box`,
which basically takes whatever `type` currently is and adds "box".

Since the list command uses `src` backups to list valid labels, we use `awk` to query the
list of filenames provided by Box.