feat: Add box support to `mdl restore`
By provide a `-b` or `--box` option to `mdl restore`, the script will use files on Box.com
instead of files in the local `backup` directory. However, it just downloads the Box.com
files into the local `backup` directory and then restores them as local backup files.

We also added other options:

`-r` or `--rm` will remove the local backup files after restoring them. Whereas this can
be used anytime a restore is performed, it is especially handy to remove files that were
just downloaded from Box after restoring them locally.

`-x` or `--extract` will decompress the files if they are compressed, leaving just the raw
SQL file and tar files. The restore obviously decompresses the files when restoring
anyway, but the idea here is it explicitly decompresses them so that the backup set left
behind (given you did not --rm it) will be decompressed so that subsequent restores of the
same backup set will be faster. Essentially, this is a shortcut for a common workflow that
we were doing of perhaps manually running something like `bunzip2 backup/*_mylabel_*.bz2`
after downloading the latest backup set.

Notes:

To add this functionality, we effectively change the way we initially identify what the
target backup files are. Now, we get a list of files (either locally or from Box.com) and
we determine available labels and exact target backup filenames based on the exact
filenames in the list. This is more precise and adaptive to having different sources that
the backup files originate from.

Also slightly improved the output describing the progress of the restore.