feat: Add `decompress` function to utility functions
The `decompress` function will be useful to point at any file path and just decompress it.
It will work with bzip2, gzip, and xz formats. If none of those extensions are found at
the end of the file path, it will error/abort. After successful decompression, it outputs
the new path of the file and removes the original.