feat: Abort script if not running as superuser on Linux
On Linux systems, there is an expectation that the environment will be set up as superuser
so that containers function properly. If a user attempts to run commands as a normal user,
certain parts of the script will throw errors, potentially having mixed results.

By running a check in the `common.sh` script which is loaded by all scripts, it is a
simple way to ensure all scripts will only run as superuser.