fix: Do not enforce superuser check for help pages
In common.sh, a check aborts the script if running on Linux and the user is not a
super user. However, this causes the docs build, which does not run as a super user, to
fail when it tries to pull all the script docs.

So, we add an exception that only aborts if the help flag isn't used.