fix: In restore script, do not delete SQL file when restoring
The `decompress` utility function will delete the original file when decompressing a
compressed file. This is normally what we want. But if `mdl restore` is trying to restore
a compressed SQL file, but it did not specify `--rm`, we don't expect the file to be
deleted.

Added a `--keep` option to the `decompress` function so that it will not delete the
archive in this situation. Then set `mdl restore` to utilize the `--keep` option. Note, it
will still delete the file if you specify `--rm` like expected, because it does that
explicitly down in line 136.