feat: Calculate MariaDB/Moodle versions and ports to use
Instead of calc-docker-compose-path.sh pointing to very specific files that use image
versions, we calculate the exact MariaDB and Moodle image version to use, and pass that
into a single docker-compose.yml file. This makes the implementation simpler and yet more
precise for each Moodle version.

Additionally, since Moodle environments need to be running at the same time and we may
need to access them via a system-wide Caddy server instead of a Docker-configured proxy
server, we need to assign the Moodle services consistent port numbers. (Note this isn't
necessarily required for production, but is preferable in staging environments.)

We could just require a variable in the .env file. But that is one more REQUIRED config
change. Instead, we calculate a port number based on counting their position in the list
of environment directories. If a `MOODLE_PORT` variable is added to .env, it will override
the calculated port.

We use a new script to accomplish all of this, called `calc-images.sh`. It should be
called before a script calls export-env.sh, so that the user can still override any of
these settings in their .env file if they so desire.

See #70, #71, #77.