feat: Add tunnel script
This script creates an ephemeral container whose purpose is to connect to the same network
as an environment and then provide a tunnel for a specific port to one of its containers.
For instance, and as a default since it is the best use case, it will tunnel the mariadb
container's port 3306 so that you can connect to the database with a client.

It can do this not only with a local container, which is useful, but also with a remote
container by connecting via SSH and also incorporating SSH tunneling.