fix: Adjust `mdl config` to remove wrapping quotes of values
Sometimes a value in a key/value pair must be wrapped in quotes, but the quotes should not
be part of the string. In fact, if you were to `source .env` with the key/value pair that
has quotes, the variable would not have the wrapping quotes. But since `mdl config` just
loops over the content line by line, it sees the quotes, so we manually remove them in the
script to match native `source .env` behavior.