fix: Improve Box.com integration to support pagination
Box.com restricts the number of items returned when it is listing folder contents. By
default, the limit is 100, although it can be increased up to 1,000. If there are more
items, it supports traditional pagination with any supported limit number.

So, we improved the Box.com list support to handle the pagination. If listing files, we
just continue listing files until done. If outputting the full results as JSON, we embed
the entire JSON response into a `pages` array so that any script can handle iterating over
each full response as separate pages.

The upload/download actions use the JSON response, so they're modified to drill into the
`pages` array of the output.