feat: Improve `mdl box upload` to receive the target filename
Before this improvement, `mdl box upload` would only upload a file with the same name that
it has on the local filesystem. Now, you can pass as the "target" a different filename
that will be assigned when uploading the file.

This supports `mdl copy` being able to rename labels when uploading to Box.