feat: Add new commands to help with installation, init, and config
Added `mdl info` and `mdl init` which help with viewing installation settings and doing
intialization of the system or a Moodle environment.

Improved `mdl remove` to have a --sys option to remove the system installation.

Finally, added install.sh and uninstall.sh scripts that actually handle the installation
and uninstallation of the mdl binaries.