feat: Add new commands to help with installation, init, and config
Added `mdl info` and `mdl init` which help with viewing installation settings and doing
initialization of the system or a Moodle environment.

Improved `mdl remove` to have --env and --sys options to remove the system installation or
a particular environment.

Finally, added install.sh and uninstall.sh scripts that actually handle the installation
and uninstallation of the mdl binaries.

As part of the niceties with this, we also add title to the general `mdl` tool output when
it doesn't receive a valid subcommand, and we also add a version number to the title.