feat: Add initial Moodle installation in `mdl init`
Up until now, initialization of an environment just sets it up to be able to have a backup
be restored to it. This improvement actually sets up a fresh Moodle installation in that
environment. If you know you'll just be restoring into the environment, and don't need the
fresh install, you can use --skip-install to skip this new feature.

The installation is accomplished like this:
- Ask user to select version of Moodle to install. This ultimately chooses which Git
  branch and Moodle image is used.
- Start the service. The bitnami moodle image will create the database and set up the
  environment on startup.
- Tweak configs, namely, remove the if logic for configuration of `wwwroot`.
- Add the git repo to the installed source code, since the default Moodle installation
  doesn't include the git repo, but we want to use it for upgrades. Optimize the repo if
  requested.
- Set file permissions accordingly.

Closes #3.