feat: Get user pref for Docker/Podman when setting configs in `mdl init`
Enhanced `mdl init` to look at existence of Docker or Podman on the system, and asks the
user what to do accordingly.  Handles all possible iterations:
- Both are installed
- One of them is installed
- Neither are installed

If selecting Podman, it also makes sure podman-compose is installed.

If at any time the tools aren't installed, or the user doesn't agree to use them, we just
abort the install and explain.

See #4.