feat: Add a library for basic data formatting in pretty, ini, or json
The pretty_line and pretty_bool functions are useful when outputting data in a pretty
human-readable format.

The make_ini and make_json functions, and their corresponding array functions, are simple
functions to format a collection of variables into ini or json format. For JSON, we could
have used jq for more robust/complex formatting support, but it was significantly more
complex to implement. So, for now, we just use these simpler functions.