feat: Improve `mdl info` to report on env info and to output ini or json
Just as it can report on the overall system information when you just execute `mdl info`,
now it will provide environment-specific information if you provide the environment. And
if you provide a specific field, it will conveniently return just that field for the given
environment.

Additionally, the --format option allows support for "ini" or "json" formatting. The "ini"
format is the typical `key=value` pair that is common in config files.

These features will pave the way for any future functionality that needs to do environment
introspection, and will be a useful addition for users.