refactor: Incorporate fastdb support into backup/restore scripts
Originally, it was not desirable to wrap fastdb logic into the official backup/restore
scripts, since fastdb is an unsafe backup approach. However, with the change where the DB
restore now occurs immediately as part of the restore (instead of just decompressing a SQL
dump file for when the container starts), we lose the ability to utilize fastdb to
expedite a restore in development.

By incorporating fastdb into the backup/restore scripts, we can reestablish the DX feature
of being able to rapidly restore an environment.

Example usage:

Make a normal backup with a fastdb archive, with no compression:
mdl backup mymoodle container -l wow --fastdb -c none

Restore with fastdb:
mdl restore mymoodle wow --fastdb