fix: Improve `mdl install-plugin` to use version.php for component name
Whereas it is good practice to use your component name as the output directory of a zipped
plugin, so plugin authors do not use the exact frankenstyle component name. However, they
do use the correct frankenstyle name in version.php because it is required.

To make `mdl install-plugin` more compatible, we update it to inspect version.php and get
the component name (and thus type and install path) from there.