fix: Improve environment and label name edge cases
The rule for Moodle and Docker names: Must consist only of lowercase alphanumeric
characters, hyphens, and underscores as well as start with a letter or number.

We discovered two issues:
- If a name had invalid characters, the project will fail to start up.
- If a name has an underscore (which is valid), backup operations are buggy.

To fix this, we add name validation in `mdl init`, and we improve spots where labels are
calculated for backup handling. Instead of referencing the second or third item when
separating the name into a list by underscores, we remove the start and end parts of the
name, leaving just the full label. We can also combine some of the logic into a single awk
command to make it cleaner.

Fixes #9.