fix: Use `command -v` instead of `which` for command existence checks
Recently, macOS changed its implementation of `which`, causing it to return its error
message on stdout instead of stderr when a command is not found. This breaks conventional
usage of `which` in scripts to check for command existence.

To address this, we replace all instances of `which <command>` with `command -v <command>`
which is more reliable and consistent across different Unix-like systems.