feat: Speed up and streamline backup/restore process
Use standard `docker cp` tooling to transfer tar archives of the volumes, instead of
running tar within the container. Also, if pbzip2, pigz, or pixz exist on the system, they
will be used, instead of their standard compression tool equivalents that are only
single-threaded (bzip2, gzip, xz).

If bsdtar is available on the system, it will use that to filter the tar stream of the
paths that don't need to be included in the backup. However, if bsdtar is not available,
we also handle removing these paths during the restore process as well. So, we always
ensure the caches and sessions are eliminated during restore.

Closes #6.