feat: Add support for custom compose files
Whereas the default.yml compose file will be used if an environment does not specify a
custom configuration with the `COMPOSE_FILE` variable, this feature improves the
`mdl calc-compose-path` script to actually do some work: It will check the existence of
the compose file path (first relative to the `MDL_COMPOSE_DIR` directory, then as an
absolute path), and it will output the path of the compose file, and throw an error if the
compose file does not exist.

This will enable users to customize their compose files.

Closes #14.